Fix for skipped turn on load in MP issue
After loading an MP game the first turns of human players can often be
skipped, meaning the AI gets two turns in a row. I traced the problem
down to being a result of the players erroneously being flagged as
having already submitted their ready message for the turn. The fix isn't
really a fix as I believe the problem is in the host. Instead I just
check for the problem upon the first activation of the local player and
send the unready message if need be.