Fixed iteration over AD deal request notifications
A couple of errors in CvDiplomacyRequests::CheckRemainingNotifications()
which could result CTDs, the wrong details used in deal withdrawl
notifications, or deals could be skipped over. This is due to the
erasing during processing which also assigns back to the iterator before
it has finished using it. The iterator would then be pointing to the
next element, which could even be end(). Also the iterator essentially
being incremented twice but would be processed on the next call to the
function and not be a real problem in reality.