Hack to prevent CTD when ending invalid mission
Address symptoms of some bug which can cause a CTD when calling
GameplayUnitMissionEnd for a mission that probably wasn't started. Not
sure how correct it is but I just check for a non-empty mission queue
before attempting to end the mission. It may result in more issues and
wasn't tested much.

This crash seemed to arise when a unit was for some reason found on top
of an active barbarian camp. I could not figure out exactly what had
happened.