Fix for issues from AI initiated wars in MP
Wars arising from "taunting" AI during diplomacy were not having net
messages sent in network MP. Therefor other players were not getting
notifications and desyncs were occuring in simultaneous MP.

For whatever reason the existing GameplayWarStateChanged message wasn't
having an immediate effect. The solution was to send ChangeWar message
directly. Unfortunately the message implies the sender is the aggressor.
Since the AI cannot send the message (as far as I can tell) I sent the
team id as a negative and flip around the aggressor when the message is
being responded to so everything looks OK.

The code is disabled when either warring team has more than 1 member.
Without the condition, it may work for larger teams as implemented by it
was not tested.