Fixed bug where overwritten saves save differently
There is deliberate Firaxis code which means manual multiplayer savegames are only ever copies of the most recent autosave. Overwriting save games was going through a code path that avoided the checks that enforced this. This has been fixed and overwritten behaviour is consistent with the normal behaviour.

This bug is the reason why people have been abel to avoid the "double ai turn" problem. However, it is quite possible that this was having other unwanted effects on the game when loaded again. I have not witnessed any problems but I suspect they are possible and Firaxis certainly thought so themselves, hence the checks! I believe Firaxis originally only had autosaves in MP and this is was their quick fix. They could have easily not put in the checks so there is almost certainly some perceived issues with creating fresh saves mid-turn.

Unfortunately, some people may be bothered by not having this gleature.