Fixed invisible units after stacked jump
Units could become invisible (but with icon still present) after a jumpToNearestValidPlot* from a plot with stacked units.

The fix was to allow updating of the center unit during a jump. The previous code explicitly requested that this not happen, so hopefully no bugs were (re)introduced.

The instance I tested with was when 2 barbs ended up stacked at the end of updateMoves and one needed to be moved off the plot. A explicit call to updateCenterUnit for both affected plots worked but I figure that the problem is more widespread.