Hack to prevent CTD when ending invalid mission
Address symptoms of some bug which can cause a CTD when calling
GameplayUnitMissionEnd for a mission that probably wasn't started. Not
sure how correct it is but I just check for a non-empty mission queue
before attempting to end the mission. This could leave the game in a bad state but I think I have seen similar kinds of conditions elsewhere. Due to this, I took the liberty of adding a notification to the user to warn them of potential errors but this may go against your philosophy.

This crash seemed to arise when a unit was for some reason found on top
of an active barbarian camp. I could not figure out exactly what had
happened.