Hack to prevent CTD when ending invalid mission
Address symptoms of some bug which can cause a CTD when calling
GameplayUnitMissionEnd for a mission that probably wasn't started. Not
sure how correct it is but I just check for a non-empty mission queue
before attempting to end the mission. This could leave the game in a bad state but I think I have seen similar kinds of conditions elsewhere.

This crash seemed to arise when a unit was for some reason found on top
of an active barbarian camp. I could not figure out exactly what had
happened. I am thinking maybe it spawned as the unit was automoving onto
it or some such.