fix double change of invisible visibility count for always invisible
Haven't seen any ill effects but the code was changing the visibility count twice for always invisibly units. Depending on how "always see invisibile" works, this could possibly cause visibility issues (which may have been reported). Or it could be completely harmless since it always increases as much as it decreases and only a zero value matters. I don't know the code but it just looked like a bug!