Added option to keep civ details unknown in pregame
If enabled, details of civs that have not been met by the local player will be hidden on the Staging Room screen. This is in keeping with the details of unmet civs not showing during the game itself.

Had to appropriate an existing PreGame function and give it a (nasty and confusing) second purpose in order to be able to get the Lua to query the known civs.

Not sure where the Lua goes, sorry.