Enabled Events in MP
Added code to send Event related messages to remote players. However, this is done via the foul means of hijacking the DiplomacyFromUI message and giving it a dual purpose.

Seems to work but maybe some events are problematic as I wouldn't have encountered all variations of mechanisms in limited testing.