fix skipped call to UpdateCityThreatCriteria
CvPlayer::UpdateCityThreatCriteria doesn't get called in CvPlayer::doTurn if about to block for input. Unlike the similiarly located call to do CvPlayer::doTurnPostDiplomacy, it wasn't getting called when finished blocking.

I don't really know what the call does or if this is an actual problem but it certainly seemed like an oversight.