fix processPlayerAutoMoves initialisation issue.
Since the var is a static, it is only initialised once (to false) and I think it is possible that it could be start of as true if a game is exited and another started. It is possible that this could never result in any issue but it is easier to remove the potiential than step through it all.

Pretty lazy fix. Probably should be a member instead of static.