fixed wrong or missing civ icons in stagingroom
Mysterious bug where the civ icons in the MP staging room would be wrong or missing.

Seemed to be a caching issue and may be more prevailent than this. Did not determine root cause but could be an issue with SimpleCivIconHookup. Fixed using the brute method of forcing a reload of the StagingRoom UI before prior to every time it is opened.

Can be reproduced by loading an MP game, clicking from the Staging Room and creating a new game with different civs playing.