fix runtime error in mplist.lua with observer slots
The fix is a bare minimum. Each observer will have their own "Observers" label, which seems odd but I guess that is the way it is - better than not being visible at all.

To repo, just create a MP game with observers. The observer slot may need to be empty. I had 1 human and 2 AI. The player list in the UI will not be present and there will be a error in the lua log.