fix wrong or missing civ icons in staging room
Only an issue when EUI present as it overrides the default IconHookup functions and does some caching using a playerID to civ info mapping. This mapping can change in PreGame and causes civ icons to be wrong ro missing when it does.

The fix here is simply to (quite heavy-handedly) reset the cache whenever the mapping may have changed. This was implemented in the StagingRoom code and references EUI. It doesn't need to be there for non-EUI installs but works for either.

I suppose it should really only be fixed in EUI or the EUI compatibility mod?