fix processPlayerAutoMoves initialisation issue.
Used to be a static in updateMoves but made a member due to it not being re-inited and maybe causing issues when loading after an exit-to-menu or maybe. Not serialized - I wasn't willing to think about the implications.

If the static var started off true I think it may have been possible that automoves might processed before they should have been or even not at all. I don't full understand the turn process but the code seems to think it important to not start before everyone is ready.

Could this cause rolling desyncs in human only MP games? It seems it would have been a big problem if that was the case though.