Preferring GetOtherPlayer over GetFromPlayer
Playing it safe, should be OK to use pDeal->GetFromPlayer() but code was using GetActivePlayer so maybe the From field wasn't always the human (although in my testing it was fine!). The context suggests that the human *should* always be From player but the human will always be Other player (unless they are both AI but that is another story).