removed odd warning disables
A number of pragmas were suppressing warnings for dereferencing null pointers. The surrounding comments didn't make sense (to me) and in some cases nearby code was dereferencing those very pointers without any assignments in between.

Might break something since the code was written for a reason, although it might not be a good one.

Better to know about an error. I want to believe the original
programmers know something I don't and these are necessary.